requirements.txt
asgiref==3.5.2
asttokens==2.2.1
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
colorama==0.4.6
comm==0.1.2
debugpy==1.6.5
decorator==5.1.1
distlib==0.3.5
Django==4.1
django-ckeditor==6.5.1
django-js-asset==2.0.0
djangorestframework==3.13.1
docopt==0.6.2
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.8.0
idna==3.3
ipykernel==6.20.2
ipython==8.8.0
jedi==0.18.2
jupyter_client==7.4.9
jupyter_core==5.1.3
matplotlib-inline==0.1.6
mysqlclient==2.1.1
nest-asyncio==1.5.6
numpy==1.23.2
openpyxl==3.0.10
packaging==21.3
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.11
platformdirs==2.5.2
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.1
pywin32==305
pyzmq==25.0.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.6.2
tornado==6.2
traitlets==5.8.1
tzdata==2022.1
urllib3==1.26.12
virtualenv==20.16.3
wcwidth==0.2.6
XlsxWriter==3.0.7
yarg==0.1.9
